Work in progress

+++ database/migrations/2017_08_30_234531_create_jobs_table.php
@@ -17,11 +17,10 @@ class CreateJobsTable extends Migration
             $table->bigIncrements('id');
             $table->string('queue');
             $table->longText('payload');
-            $table->unsignedTinyInteger('attempts');
+            $table->unsignedBigInteger('attempts');
             $table->unsignedInteger('reserved_at')->nullable();
             $table->unsignedInteger('available_at');
             $table->unsignedInteger('created_at');
-
             $table->index(['queue', 'reserved_at']);
         });
     }
